backend/requirements.txt
Flask~=3.1.0
dotenv~=0.9.9
python-dotenv~=1.1.0
openai~=1.76.0
fitz~=0.0.1.dev2
faiss-cpu~=1.11.0
tiktoken~=0.9.0


flask-cors~=5.0.1
numpy~=2.1.3